pyproject.toml
[project]
name = "event-agent"
version = "0.25.0a1"
description = "AI agent crew that scrapes, classifies, and stores tech/BI events"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "litellm>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "geopy>=2.4.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "tenacity>=9.0.0",
    "rapidfuzz>=3.10.0",
    "python-dateutil>=2.9.0",
    "playwright>=1.44.0",
    "apscheduler>=3.10.0,<4.0",
    "playwright-stealth>=2.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
